backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
prometheus-fastapi-instrumentator
psycopg2-binary
pydantic
groq
faker
httpx
pytest
pytest-cov
pytest-asyncio
